Don't create api instances in the authorizer.
Construction of API objects is deeply tied into other components of
the application, and the authorizer really only has to know how to
select a particular class of API that matches a token. Extra
indirection is icky, but this way the authorization doesn't have to
know other parts of the app need to be put together.