Automatically convert UnauthorizedExceptions to 401 responses.
A 401 response is appropriate if the credentials presented are simply
not valid for this particular operation (i.e. other credentials might
work). However, the server is supposed to respond with the same
challenge in that case; for maintenance purposes, it's easier to let
the decorator handle constructing the response in that case.