Restructure to avoid missing-method API errors.
Rather than having to check whether the request's api is of the right
sort to perform some action -- and it should be noted that due to the
capability-like nature of the api objects, this is more about avoid
missing-method errors than it is about doing an actual security check
-- all api objects have default methods for all valid operations that
raise an UnauthorizedException, which need only be overridden if the
api actually supports the operation. This way, clients can use the api
objects agnostically.